refactor: move tree and hyperplane impl
