+++ src/db/database.rs
@@ -1,1 +1,1 @@
-
+pub struct Database {}
