+++ src/ix/mod.rs
@@ -1,7 +1,12 @@
+mod hyperplane;
 mod index;
 mod tree;
 mod vector;
 
+use rand::prelude::SliceRandom;
+use std::collections::HashMap;
+
+use hyperplane::*;
 use tree::*;
 
 pub use index::*;
